Added automatic return to Jekyll site
Uses the "comment_received.html" page and requires PHP, so it is not possible to place this page on the static Jekyll site.

TODO: Currently, "return_url" will be a relative url (relative to the PHP submit page). Fix it!

Also made slight changes to display a simple error if the email did not send, just so developers aren't entirely in the dark in case comments stop working. (should that be a separate commit?)