Added "Reply to" field to mail header
This extra field will only be added if the user has supplied an email address.

This fix allows blog owners to reply to the user's comment by replying to their email (if they included an email address in the comment) if they want a private reply rather than a reply as a comment on the blog.